% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_functions.R
\name{add_gene_annotations}
\alias{add_gene_annotations}
\title{add_gene_annotations}
\usage{
add_gene_annotations(df, gene_column = "gene", org = "human", release = "105")
}
\arguments{
\item{df}{Dataframe}

\item{gene_column}{Column contaning gene symbols}

\item{org}{Organism, either "human" or "mouse". Default is "human".}

\item{release}{Ensembl release. Default is "105".}
}
\value{
Dataframe with functional annotation of genes
}
\description{
Save dataframe as GCT file
}
