% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversion.R
\name{convert_human_to_mouse}
\alias{convert_human_to_mouse}
\title{convert_human_to_mouse}
\usage{
convert_human_to_mouse(genes, one.to.many = F)
}
\arguments{
\item{genes}{A vector of human gene symbols.}

\item{one.to.many}{Enable many mouse gene symbols to mapped to same humangene}
}
\value{
A vector of unique mouse gene symbols that are mapped one-to-one. If one.to.many = TRUE, returns a dataframe that mapped one-to-many.
}
\description{
Convert human (HGNC) gene symbols to mouse (MGI) gene symbols.
}
