% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_functions.R
\name{convert_list_to_df}
\alias{convert_list_to_df}
\title{convert_list_to_df}
\usage{
convert_list_to_df(list)
}
\arguments{
\item{list}{A list of vectors}
}
\value{
A dataframe
}
\description{
Store each vector from a list into a dataframe columns
}
