% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversion.R
\name{convert_mouse_to_human}
\alias{convert_mouse_to_human}
\title{convert_mouse_to_human}
\usage{
convert_mouse_to_human(genes, one.to.many = F)
}
\arguments{
\item{genes}{A vector of mouse gene symbols.}

\item{one.to.many}{Enable many human gene symbols to mapped to same mouse gene}
}
\value{
A vector of unique human gene symbols that are mapped one-to-one. If one.to.many = TRUE, returns a dataframe that mapped one-to-many.
}
\description{
Convert mouse (MGI) gene symbols to human (HGNC) gene symbols.
}
