% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_genes.R
\name{get_mt_genes}
\alias{get_mt_genes}
\title{get_mt_genes}
\usage{
get_mt_genes(org = "human")
}
\arguments{
\item{org}{Organism either "human" or "mouse". Default is "human".}
}
\value{
A vector of MT genes.
}
\description{
Get gene symbols from mitochondrial chromosome.
}
