% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_genes.R
\name{get_str_genes}
\alias{get_str_genes}
\title{get_str_genes}
\usage{
get_str_genes(org = "human", str = c("rb"))
}
\arguments{
\item{org}{Organism either "human" or "mouse". Default is "human".}

\item{str}{String to search for. Default is "rb".}
}
\value{
A vector of str-matching genes.
}
\description{
Get gene symbols from default strings.
}
