% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_genes.R
\name{get_xy_genes}
\alias{get_xy_genes}
\title{get_xy_genes}
\usage{
get_xy_genes(org = "human")
}
\arguments{
\item{org}{Organism either "human" or "mouse". Default is "human".}
}
\value{
A vector of XY genes.
}
\description{
Get gene symbols from the XY chromosome.
}
