% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversion.R
\name{import_biomart}
\alias{import_biomart}
\title{get_biomart}
\usage{
import_biomart(host = "https://dec2021.archive.ensembl.org")
}
\arguments{
\item{host}{URL to retrieve archived ensembl biomarts}
}
\value{
biomart_dict in .GlobalEnv
}
\description{
Retrieves archived ensembl biomarts from web.
}
