% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversion.R
\name{import_biomart_local}
\alias{import_biomart_local}
\title{import_biomart_local}
\usage{
import_biomart_local(release = "105")
}
\arguments{
\item{release}{Ensembl release number}
}
\value{
biomart_dict in .GlobalEnv
}
\description{
Retrieves archived ensembl biomarts from local directory.
}
