% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_functions.R
\name{write_gct}
\alias{write_gct}
\title{write_gct}
\usage{
write_gct(df, save_name = "", save_dir = getwd())
}
\arguments{
\item{df}{Dataframe}

\item{save_name}{File name}

\item{save_dir}{File path}
}
\value{
Dataframe in GCT format
}
\description{
Save dataframe as GCT file
}
