% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_functions.R
\name{write_gmt}
\alias{write_gmt}
\title{write_gmt}
\usage{
write_gmt(df, save_name = "", save_dir = getwd())
}
\arguments{
\item{df}{Dataframe}

\item{save_name}{File name}

\item{save_dir}{File path}
}
\value{
Dataframe in GMT format
}
\description{
Save dataframe as GMT file
}
